requirements.txt
# NetBird MSP Appliance - Python Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==41.0.7
python-multipart==0.0.6
httpx==0.25.2
jinja2==3.1.2
docker==7.0.0
psutil==5.9.7
pyyaml==6.0.1
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-httpx==0.28.0
